pyproject.toml
[project]
name = "strata-fdtd"
version = "0.1.0"
description = "FDTD acoustic simulation toolchain for loudspeaker design and metamaterials"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "rwalters" }
]
keywords = ["fdtd", "acoustics", "audio", "loudspeaker", "metamaterial", "simulation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.15.3",
    "h5py>=3.8",
    "click>=8.0",
    "rich>=13.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
# GPU acceleration (requires PyTorch)
gpu = [
    "torch>=2.0",
]

# Geometry and CAD export
geometry = [
    "ezdxf>=1.0",
    "numpy-stl>=3.0",
    "scikit-image>=0.21",
]

# Native C++ kernels (optional, for performance)
native = []

# Development tools
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

# Full installation
all = [
    "strata-fdtd[gpu,geometry,dev]",
]

[project.urls]
Homepage = "https://github.com/rjwalters/strata-fdtd"
Repository = "https://github.com/rjwalters/strata-fdtd"
Documentation = "https://github.com/rjwalters/strata-fdtd/tree/main/docs"
"Bug Tracker" = "https://github.com/rjwalters/strata-fdtd/issues"

[project.scripts]
fdtd-compute = "strata_fdtd.cli.compute:main"

[build-system]
requires = [
    "scikit-build-core>=0.5",
    "pybind11>=2.11",
]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
cmake.version = ">=3.15"
cmake.build-type = "Release"
cmake.source-dir = "src/strata_fdtd/_kernels"
ninja.version = ">=1.5"
wheel.packages = ["src/strata_fdtd"]
wheel.install-dir = "strata_fdtd"

[tool.scikit-build.cmake.define]
FDTD_USE_OPENMP = "ON"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["strata_fdtd"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "benchmark: marks tests as benchmarks",
    "slow: marks tests as slow",
    "gpu: requires GPU support",
]
